requirements.txt
streamlit==1.26.0
pillow==9.5.0
